+++ client/src/assets/styles/Styles.js
@@ -2,7 +2,7 @@ const styles = (theme) => ({
   grid: {
     backgroundColor: theme.palette.primary.main,
     borderRadius: "20px",
-    padding: 15,
+    padding: 5,
     paddingLeft: 25,
   },
   textField: {
@@ -34,7 +34,7 @@ const styles = (theme) => ({
     textAlign: "left",
   },
   rowMargin: {
-    margin: 20,
+    margin: 10,
   },
 });
